// src/app/api/model/currentUserInfoDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRoleDto } from './userRoleDto';
import { UserDto } from './userDto';


export interface CurrentUserInfoDto { 
    role: UserRoleDto;
    info?: UserDto;
}