// src/app/api/model/userRoleDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum UserRoleDto {

    SUPER_ADMIN = 'SUPER_ADMIN',

    MODERATOR = 'MODERATOR',

    CONTENT_MANAGER = 'CONTENT_MANAGER',

    VIEWER = 'VIEWER',

    ANONYMOUS = 'ANONYMOUS'
}